backend/app/services/shadowban_shield.py
"""
VUC-2026 Shadowban Shield Service
YouTube bot deception and content protection
"""

import os
import random
import numpy as np
from typing import Dict, Any, Optional, Tuple
import cv2
import logging
from PIL import Image, ImageEnhance


class ShadowbanShield:
    """Shadowban protection system for YouTube content"""
    
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.logger = logging.getLogger(__name__)
        
        # Shadowban shield settings
        self.pixel_noise_enabled = config.get("pixel_noise_enabled", True)
        self.pixel_noise_intensity = config.get("pixel_noise_intensity", 0.1)
        self.speed_variation_enabled = config.get("speed_variation_enabled", True)
        self.speed_variation_range = config.get("speed_variation_range", (0.99, 1.01))
        self.frame_jitter_enabled = config.get("frame_jitter_enabled", True)
        self.frame_jitter_intensity = config.get("frame_jitter_intensity", 0.001)
    
    def apply_protection(self, video_assets: Dict[str, Any]) -> Dict[str, Any]:
        """Apply all shadowban protection measures"""
        try:
            self.logger.info("Applying shadowban shield protection")
            
            protected_assets = video_assets.copy()
            
            # Apply pixel noise if enabled
            if self.pixel_noise_enabled:
                protected_assets = self._apply_pixel_noise(protected_assets)
            
            # Apply speed variation if enabled
            if self.speed_variation_enabled:
                protected_assets = self._apply_speed_variation(protected_assets)
            
            # Apply frame jitter if enabled
            if self.frame_jitter_enabled:
                protected_assets = self._apply_frame_jitter(protected_assets)
            
            # Add metadata variations
            protected_assets = self._add_metadata_variations(protected_assets)
            
            self.logger.info("Shadowban shield protection applied successfully")
            return protected_assets
            
        except Exception as e:
            self.logger.error(f"Failed to apply shadowban shield: {e}")
            return video_assets
    
    def _apply_pixel_noise(self, assets: Dict[str, Any]) -> Dict[str, Any]:
        """Apply invisible pixel noise to frames"""
        try:
            video_path = assets.get("video_path")
            if not video_path or not os.path.exists(video_path):
                return assets
            
            # Open video
            cap = cv2.VideoCapture(video_path)
            fps = int(cap.get(cv2.CAP_PROP_FPS))
            width = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
            height = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
            total_frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
            
            # Create output video path
            output_path = video_path.replace(".", "_shielded.")
            fourcc = cv2.VideoWriter_fourcc(*'mp4v')
            out = cv2.VideoWriter(output_path, fourcc, fps, (width, height))
            
            frame_count = 0
            while True:
                ret, frame = cap.read()
                if not ret:
                    break
                
                # Apply pixel noise
                if frame_count % 30 == 0:  # Apply every 30 frames
                    noise = np.random.normal(0, self.pixel_noise_intensity, frame.shape).astype(np.uint8)
                    frame = cv2.add(frame, noise)
                
                out.write(frame)
                frame_count += 1
                
                # Progress logging
                if frame_count % 100 == 0:
                    self.logger.debug(f"Processed {frame_count}/{total_frames} frames")
            
            cap.release()
            out.release()
            
            assets["video_path"] = output_path
            assets["pixel_noise_applied"] = True
            
            return assets
            
        except Exception as e:
            self.logger.error(f"Pixel noise application failed: {e}")
            return assets
    
    def _apply_speed_variation(self, assets: Dict[str, Any]) -> Dict[str, Any]:
        """Apply micro speed variations to avoid detection"""
        try:
            video_path = assets.get("video_path")
            if not video_path:
                return assets
            
            # Generate speed variation profile
            min_speed, max_speed = self.speed_variation_range
            speed_profile = []
            
            # Create speed variations every 10 seconds
            video_duration = assets.get("duration", 300)  # Default 5 minutes
            segments = int(video_duration / 10)
            
            for i in range(segments):
                speed = random.uniform(min_speed, max_speed)
                speed_profile.append(speed)
            
            assets["speed_profile"] = speed_profile
            assets["speed_variation_applied"] = True
            
            self.logger.info(f"Applied speed variation with {len(speed_profile)} segments")
            return assets
            
        except Exception as e:
            self.logger.error(f"Speed variation application failed: {e}")
            return assets
    
    def _apply_frame_jitter(self, assets: Dict[str, Any]) -> Dict[str, Any]:
        """Apply frame jitter to simulate human editing"""
        try:
            video_path = assets.get("video_path")
            if not video_path:
                return assets
            
            # Generate frame jitter pattern
            jitter_pattern = []
            
            # Random frame drops/duplicates (very subtle)
            video_duration = assets.get("duration", 300)
            fps = assets.get("fps", 30)
            total_frames = int(video_duration * fps)
            
            for i in range(total_frames):
                if random.random() < self.frame_jitter_intensity:
                    # Either drop or duplicate frame
                    jitter_pattern.append(random.choice(["drop", "duplicate"]))
                else:
                    jitter_pattern.append("normal")
            
            assets["frame_jitter_pattern"] = jitter_pattern
            assets["frame_jitter_applied"] = True
            
            self.logger.info(f"Applied frame jitter to {len(jitter_pattern)} frames")
            return assets
            
        except Exception as e:
            self.logger.error(f"Frame jitter application failed: {e}")
            return assets
    
    def _add_metadata_variations(self, assets: Dict[str, Any]) -> Dict[str, Any]:
        """Add subtle metadata variations"""
        try:
            # Vary thumbnail slightly
            thumbnail_path = assets.get("thumbnail_path")
            if thumbnail_path and os.path.exists(thumbnail_path):
                varied_thumbnail = self._vary_thumbnail(thumbnail_path)
                assets["thumbnail_path"] = varied_thumbnail
            
            # Add metadata timestamp variations
            import datetime
            now = datetime.datetime.utcnow()
            
            # Add random seconds to upload time
            random_seconds = random.randint(1, 59)
            upload_time = now + datetime.timedelta(seconds=random_seconds)
            
            assets["scheduled_upload_time"] = upload_time.isoformat()
            assets["metadata_variations_applied"] = True
            
            return assets
            
        except Exception as e:
            self.logger.error(f"Metadata variations failed: {e}")
            return assets
    
    def _vary_thumbnail(self, thumbnail_path: str) -> str:
        """Apply subtle variations to thumbnail"""
        try:
            # Open image
            img = Image.open(thumbnail_path)
            
            # Apply very subtle brightness variation
            enhancer = ImageEnhance.Brightness(img)
            brightness_factor = random.uniform(0.98, 1.02)  # ±2% variation
            varied_img = enhancer.enhance(brightness_factor)
            
            # Apply very subtle contrast variation
            enhancer = ImageEnhance.Contrast(varied_img)
            contrast_factor = random.uniform(0.98, 1.02)  # ±2% variation
            varied_img = enhancer.enhance(contrast_factor)
            
            # Save varied thumbnail
            varied_path = thumbnail_path.replace(".", "_varied.")
            varied_img.save(varied_path)
            
            return varied_path
            
        except Exception as e:
            self.logger.error(f"Thumbnail variation failed: {e}")
            return thumbnail_path
    
    def generate_upload_schedule(self, video_count: int) -> list:
        """Generate optimized upload schedule with jitter"""
        try:
            schedule = []
            base_interval = 3600  # 1 hour base interval
            
            for i in range(video_count):
                # Add ±15% jitter to upload timing
                jitter = random.uniform(0.85, 1.15)
                interval = base_interval * jitter
                
                # Random time of day preference (avoid peak hours)
                hour = random.choice([9, 10, 11, 14, 15, 16, 17, 18, 19, 20])
                minute = random.randint(0, 59)
                
                schedule.append({
                    "video_index": i,
                    "preferred_hour": hour,
                    "preferred_minute": minute,
                    "interval_seconds": interval,
                    "jitter_applied": True
                })
            
            return schedule
            
        except Exception as e:
            self.logger.error(f"Upload schedule generation failed: {e}")
            return []
    
    def validate_content_uniqueness(self, video_path: str) -> Dict[str, Any]:
        """Validate content uniqueness to avoid duplicate detection"""
        try:
            # Placeholder for content uniqueness validation
            # In production, this would use perceptual hashing or similar
            
            validation_result = {
                "is_unique": True,
                "similarity_score": 0.15,  # Low similarity
                "duplicate_detected": False,
                "validation_passed": True
            }
            
            return validation_result
            
        except Exception as e:
            self.logger.error(f"Content uniqueness validation failed: {e}")
            return {"validation_passed": False, "error": str(e)}
