backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
yt-dlp==2023.12.30
google-generativeai==0.3.2
google-genai==1.15.0
google-api-python-client==2.108.0
google-cloud-aiplatform==1.38.0
google-cloud-bigtable==2.19.0
google-cloud-bigquery==3.13.0
google-cloud-bigquery-datatransfer==3.9.0
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
google-cloud-monitoring==2.17.1
google-cloud-pubsub==2.19.0
google-cloud-compute==1.14.1
google-cloud-sql==1.34.0
google-cloud-redis==2.14.0
google-cloud-spanner==3.9.0
google-cloud-dataflow==2.13.0
google-cloud-ml==1.4.0
google-cloud-container==2.14.0
google-cloud-run==0.10.0
google-cloud-functions==1.13.0
google-cloud-secret-manager==2.16.0
google-cloud-artifact-registry==1.9.0
google-cloud-build==1.11.0
google-cloud-scheduler==2.13.0
google-cloud-tasks==2.14.0
google-cloud-dns==0.34.0
google-cloud-filestore==1.15.0
google-cloud-network-management==1.12.0
google-cloud-web-security-scanner==1.6.0
google-cloud-resource-manager==1.10.0
google-auth==2.23.4
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
edge-tts==6.1.10
ffmpeg-python==0.2.0
pillow==10.1.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.28.1
aiofiles==23.2.1
python-dotenv==1.0.0
psutil==5.9.7
mediapipe==0.10.11
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
exifread==3.0.0
sentence-transformers==2.2.2
pgvector==0.2.3
grpcio==1.59.3
grpcio-tools==1.59.3
proto-plus==1.22.3
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4
joblib==1.3.2
aiohttp==3.9.1
mimetypes==1.0.0
turkish-string==1.0.0

# Connection Resilience Dependencies
backoff==2.2.1
tenacity==8.2.3
circuitbreaker==1.4.0
dnspython==2.4.2
ping3==4.0.4
aiodns==3.1.1
